+++ api/src/main/java/de/maxhenkel/voicechat/api/events/ClientEvent.java
@@ -1,0 +1,12 @@
+package de.maxhenkel.voicechat.api.events;
+
+import de.maxhenkel.voicechat.api.VoicechatClientApi;
+
+public interface ClientEvent extends Event {
+
+    /**
+     * @return the voice chat client API
+     */
+    VoicechatClientApi getVoicechat();
+
+}