+++ common/src/main/java/de/maxhenkel/voicechat/plugins/impl/VoicechatClientApiImpl.java
@@ -1,0 +1,7 @@
+package de.maxhenkel.voicechat.plugins.impl;
+
+import de.maxhenkel.voicechat.api.VoicechatClientApi;
+
+public class VoicechatClientApiImpl extends VoicechatApiImpl implements VoicechatClientApi {
+
+}